REFACTOR: cmd directory
